refactor - #145 : 내정보 수정 - MyPageTab 삭제
